refactor: MovieScheduler & DateScheduler 삭제 및 TimeScheduler Object로 변경
