--- app/src/main/java/woowacourse/movie/domain/MovieScheduler.kt
@@ -1,6 +1,0 @@
-package woowacourse.movie.domain
-
-class MovieScheduler(
-    val dateScheduler: DateScheduler,
-    val timeScheduler: TimeScheduler,
-)
